:ambulance: Fix critical bug. - Fixed critical bug when use models and schemas.
